refactored products and reviews stores
