--- client/reducers/index.js
@@ -1,12 +1,0 @@
-import { productReducer, getProducts } from './productReducer'
-import authReducer from './authReducer'
-import { reviewReducer, createReview, getAllReviews } from './reviewReducer'
-
-export { 
-  productReducer,
-  getProducts,
-  authReducer,
-  reviewReducer,
-  createReview,
-  getAllReviews
-}
