+++ client/store/thunks.js
@@ -1,0 +1,31 @@
+import axios from 'axios';
+
+import { _getProduct } from './actionCreators';
+
+export const getProducts = () => {
+  return (dispatch) => {
+    return axios.get('/api/products')
+      .then(response => response.data)
+      .then(products => dispatch(_getProducts(products)))
+      .catch(error => console.log(error.message))
+  }
+}
+
+export const getAllReviews = (reviews) => {
+  return (dispatch) => {
+    axios.get('/api/reviews')
+      .then(response => response.data)
+      .then(reviews => dispatch(_getAllReviews(reviews)))
+      .catch(error => console.log(error.message))
+  }
+}
+
+
+export const createReview = (review) => {
+  return (dispatch) => {
+    axios.post(`/api/products/${id}/reviews`, review)
+      .then(response => response.data)
+      .then(review => dispatch(_createReview(review)))
+      .catch(error => console.log(error.message))
+  }
+}
